+++ changelog.md
@@ -6,3 +6,7 @@
 
 # 1.0.2
 - Code rework/cleanup
+
+# 1.1.0
+- Compatibility for More Icons by hiimjustin000
+- Compatilibity for Seperate Dual Icons by Weebify